requirement.txt
asttokens==3.0.0
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.3
cryptography==45.0.7
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.0.2
RapidFuzz==3.14.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
et_xmlfile==2.0.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
xlrd==2.0.1